FEATURE: Keep track of message timestamp when it's merged (#94)
Currently when similar messages are grouped, we lose the timestampes of
all the messages that are merged except for the most recent message's
timestamp. What this PR does is it stores the timestamp of the message
to be merged in its env hash so you can see the individual timestamp of
each merged message.