PERF: Defer sending env to client until it's needed
Currently when you go to Logster homepage, the server fetches the last
50 messages with their envs and send them to your browser. 50 messages
with their envs can be 5+ MB in size. This commit changes the server so
now it'll only serve the messages without their envs, and when you
select a message and switch to the env tab, the client will request the
env associated with the message you selected via the new endpoint
`/fetch-env/${msg.key}.json`.