FEATURE: Linkify backtrace lines to github (#104)
This allows Logster to construct GitHub URLs for each backtrace line based on a config provided by you that maps paths on the file system to GitHub repository URLs.

The config for a Rails app may look like this:

```ruby
Logster.config.project_directories = [
  { path: Rails.root.to_s, url: "https://github.com/<your_org>/<your_repo>" }
]
```

The GitHub links will point to the `master` branch. Add a boolean key `main_app: true` to the hash to instruct Logster to use the `application_version` attribute from the env tab when constructing the links.