FIX: Fetch env when row is selected to allow copy to work on Firefox
The copy button we recently introduced needs some information from the
message's env, however we only fetch the env when the user selects the
message AND if the user has the env tab selected. My solution to this
was to fetch the env when the user clicked the copy button and then
write to the clipboard.

This works on Chrome and IE11, but not on Firefox because Firefox
requires `document.execCommand('copy')` to be called inside a
user-generated event so calling that method inside `then` after fetching
env asynchronously doesn't work.