FEATURE: Improve support for logging error objects (#153)
Currently, if you rescue an exception and log the exception object to Logster, the backtrace points to the caller location rather the original location that raised the exception. This commit improves adds special handling for when an exception object is passed to Logster so that the backtrace of the exception object is used rather than the caller location.