DEV: Avoid deprecation warning in Redis 4.8
In Redis 4.8

```
Deprecate sadd and srem returning a boolean when called with a single argument. To enable the redis 5.0 behavior you can set Redis.sadd_returns_boolean = true.
```

However, we want Logster to remain backwards compatible with older Redis
version so we pass in an array instead of using `sadd?` or `srem?`.