FEATURE: Stop grouping logs with different messages (#217)
This was intended to group logs which differed only by some number or hash. However, grouping at this low-level causes the 'message' data for subsequent logs to be completely thrown away, which can be very frustrating when debugging certain problems.

This log-level grouping was originally introduced in 8e067d92, which was added before we had UI-based options for adding custom grouping patterns. Nowadays, those custom grouping patterns are a much better option, and preserve the original message of each log