FIX: Add compatibility with Rails 7.1+
Rails 7.1 introduced a new default logger
(`ActiveSupport::BroadcastLogger`) that responds to a new method named
`#broadcast_to`.

Our current logger isn’t compatible with it and we’re overwriting the
Rails logger. This means that a Rails 7.1 will call `#broadcast_to` on
the logger which will break.

This patch fixes the issue by using the new broadcast logger and adding
our Logster logger to it when inside a Rails 7.1+ app.

For Rails < 7.1, we continue to replace the Rails logger with the
Logster one.