FEATURE: Update `Logster::Logger#chain` method to accept `with_opts`
This commit updates `Logster::Logger#chain` method to accept a
`with_opts` keyword argument which when set to `true` will call the
chained logger's `add` methood with an extra `opts` parameter which is a
hash consistenting of Logster's env and other useful information like the
backtrace of an exception.

This improves `Logster::Logger#chain` API so that other loggers can
benefit from the envs and backtraces that Logster has already
painstakingly gathered.