FIX: Don't break search when providing invalid regex characters
Searching through messages accepts terms as a regexp or a string.
When a regexp is provided, it is parsed and checked, and if it’s
invalid, then it’s discarded, without raising an error.
But when a string is provided, we convert it to a regexp as-is and if it
results in an invalid regexp, then an error is raised.

This patch addresses that issue by escaping the string before converting
it to a regexp, that way it can’t create an invalid regexp anymore.